[IMP] runbot_merge: move batching tests out of test_basic
